+++ src/WireChatServiceProvider.php
@@ -14,9 +14,9 @@ use Namu\WireChat\Livewire\Components\NewChat;
 use Namu\WireChat\Livewire\Components\NewGroup;
 use Namu\WireChat\Livewire\Group\Permissions;
 use Namu\WireChat\Livewire\Index;
-use Namu\WireChat\Livewire\Info\AddMembers;
+use Namu\WireChat\Livewire\Group\AddMembers;
 use Namu\WireChat\Livewire\Info\Info;
-use Namu\WireChat\Livewire\Info\Members;
+use Namu\WireChat\Livewire\Group\Members;
 use Namu\WireChat\Livewire\Modals\ChatDialog;
 use Namu\WireChat\Livewire\Chat\Drawer;
 use Namu\WireChat\Livewire\View;
